# Optimize encode by avoiding list of field props